working on it

+++ autoupload.py
@@ -2,7 +2,7 @@
 import datetime
 import json
 import os.path
-
+from pathlib import Path
 
 import upload_youtube
 import config
@@ -23,13 +23,53 @@ def save_meta(meta_file, meta):
         json.dump(meta, fh)
 
 
+def get_tmp_file(file):
+    return(config.tmp_dir+os.path.basename(file))
+
+
+def get_tag_file(file, tag):
+    return (file+"."+tag)
+
+def tag_file(file, tag):
+    Path(get_tag_file(file,tag)).touch()
+
+def is_tagged(file, tag):
+    return os.path.isfile(get_tag_file(file,tag))
+
 def concat_videos(video_files, output):
+
+    print ("Concatting {} to {}".format(video_files, output))
+
+    if os.path.isfile(output):
+        print ("Already concatted")
+        return
+
+
     list_file_name = config.upload_dir + "files.txt"
     with open(list_file_name, "w") as fh:
         for video_file in video_files:
             fh.write("file '{}'\n".format(video_file))
 
-    subprocess.check_call("ffmpeg -y -f concat -safe 0 -i '{}' -c copy '{}'".format(list_file_name, output), shell=True)
+    subprocess.check_call("ffmpeg -y -f concat -safe 0 -i '{}' -c copy '{}'".format(list_file_name, get_tmp_file(output)), shell=True)
+    os.rename(get_tmp_file(output), output)
+
+def do_uploads(video_file, meta):
+    print ("Processing uploads for {}".format(video_file))
+
+    youtube_done_file=video_file+".youtube"
+    if not is_tagged(video_file, 'youtube'):
+        print("Uploading to youtube")
+        upload_youtube.upload(
+            video_file,
+            "My cats eating {}".format(meta['date']),
+            "Every time a cat eats it streams automaticly to youtube.\nIts a scale and automated feeder i build to diet my cat. (Mogwai, the black white one)\nThe grey cat is Tracy. :)\n\nMore info: https://github.com/psy0rz/meowton/wiki",
+            ["cats", "animal", "eating", "food", "meowton"],
+            "animals",
+            "en-US",
+            "public"
+        )
+        tag_file(video_file, "youtube")
+        print("Done")
 
 
 print("Scanning " + config.upload_dir)
@@ -54,14 +94,9 @@ for meta_file in meta_files:
 
     if day != prev_day:
         # upload
-        output_file="{}My cats eating, {}.h264".format(config.upload_dir, prev_day)
-        concat_videos(day_files, output_file)
-        # if not 'youtube' in meta['done']:
-        #     print("Uploading to youtube")
-        #     # upload_youtube.upload(video_file,**meta)
-        #     # meta['done'].append('youtube')
-        #     save_meta(meta_file, meta)
-        #     print("Done")
+        concat_file="{}{}.h264".format(config.upload_dir, prev_day)
+        concat_videos(day_files, concat_file)
+        do_uploads(concat_file, { "date": prev_day})
 
         prev_day = day
         day_files = []
